Conform `doobie-munit` with `doobie-specs2`
