+++ modules/munit/src/main/scala/doobie/munit/package.scala
@@ -1,0 +1,12 @@
+// Copyright (c) 2013-2020 Rob Norris and Contributors
+// This software is licensed under the MIT License (MIT).
+// For more information see LICENSE or https://opensource.org/licenses/MIT
+
+package doobie
+
+package object munit {
+
+  type Checker[M[_]] = analysisspec.Checker[M]
+  type IOChecker     = analysisspec.IOChecker
+
+}
